Refactor : remove Validator class
